moved core e2e to new space
