--- core/tests/test_core_service.py
@@ -1,42 +1,0 @@
-import pytest
-
-from core import services as core_services
-from user.exceptions import UserExists
-
-
-@pytest.mark.parametrize("basic_user", [None])
-def test_create_identity(mocker):
-    mock_user_create = mocker.patch("user.services.create")
-    mock_profiles_create = mocker.patch(
-        "profiles.services.create_from_sso",
-        return_value="__profile__",
-    )
-    # User is created
-    profile = core_services.create_identity(
-        id="john.sso.email.id@gov.uk",
-        first_name="Billy",
-        last_name="Bob",
-        all_emails=[
-            "test@test.com",
-            "test2@test.com",
-            "test3@test.com",
-        ],
-        primary_email="test2@test.com",
-        contact_email="test3@test.com",
-    )
-    mock_user_create.assert_called_once_with(
-        sso_email_id="john.sso.email.id@gov.uk",
-    )
-    mock_profiles_create.assert_called_once_with(
-        sso_email_id="john.sso.email.id@gov.uk",
-        first_name="Billy",
-        last_name="Bob",
-        all_emails=[
-            "test@test.com",
-            "test2@test.com",
-            "test3@test.com",
-        ],
-        primary_email="test2@test.com",
-        contact_email="test3@test.com",
-    )
-    assert profile == "__profile__"
